+++ reflectable/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.0.11
+
+* Fix bug 291 (such that metadata on enum values can be obtained).
+* Update dependencies to use analyzer 6.11.0 and lints 5.1.0.
+* Change the implementation to use `..._obvious_local_variable_types`
+  lints.
+* Update the [capability design document][1] slightly.
+
 ## 4.0.10
 
 * Reintroduce support for newer versions of the analyzer (6.7.0 and up).